aeromind/requirements.txt
google-genai
google-cloud-aiplatform
google-cloud-storage
google-cloud-firestore
google-adk
gqlalchemy
neo4j
xgboost
scikit-learn
pandas
numpy
scipy
fastapi
uvicorn
websockets
httpx
python-dotenv
aiohttp
aiofiles
pytest
pytest-asyncio
